added editing capabilities

+++ middleware/relocate-files.js
@@ -9,32 +9,39 @@ const path = require('path');
 const tempDirectory = './temp';
 
 module.exports = (req, res, next) => {
-    console.log('relocating files');
 
-    const projectMediaDir = './public/images/project-media';
+    if (fs.existsSync(tempDirectory)) {
+        console.log('relocating files');
 
-    const currentProjectDir = req.body.slug;
-    const finalDestination = projectMediaDir + '/' + currentProjectDir;
+        // Generate URL for file destination from media directory and current project slug
+        const projectMediaDir = './public/images/project-media';
+        const currentProjectDir = req.body.slug;
+        const finalDestination = projectMediaDir + '/' + currentProjectDir;
 
-    // Create directory for final destination
-    if (!fs.existsSync(finalDestination)) {
-        fs.mkdirSync(finalDestination, { recursive: true });
+        // Create directory for final destination
+        if (!fs.existsSync(finalDestination)) {
+            fs.mkdirSync(finalDestination, { recursive: true });
+        }
+
+        // Get array of files in temp directory
+        const files = fs.readdirSync(tempDirectory);
+
+        // Copy files to final destination
+        files.forEach(file => {
+            const oldFile = path.join(tempDirectory, file);
+            const newFile = path.join(finalDestination, file);
+            // console.log('old dest', oldFile);
+            // console.log('new dest', newFile);
+            fs.renameSync(oldFile, newFile)
+        })
+
+        // Delete temp directory
+        fs.rmdirSync(tempDirectory)
+        console.log('created new directory for files');
+        
     }
-
-    // Get array of files in temp directory
-    const files = fs.readdirSync(tempDirectory);
-    
-    // Copy files to final destination
-    files.forEach(file => {
-        const oldFile = path.join(tempDirectory, file);
-        const newFile = path.join(finalDestination, file);
-        // console.log('old dest', oldFile);
-        // console.log('new dest', newFile);
-        fs.renameSync(oldFile, newFile)
-    })
-
-    // Delete temp directory
-    fs.rmdirSync(tempDirectory)
-    console.log('created new directory for files');
-    res.json(req.body);
+    else {
+        console.log('No files to relocate.')
+    }
+    return res.json(req.body);
 }
